+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 1. [](#new)
     * Requires **Grav 1.7.11**
     * Support running Grav outside webroot [#29](https://github.com/getgrav/grav-plugin-problems/pull/29)
-    * Check whether `user/accounts` is writable [#25](https://github.com/getgrav/grav-plugin-problems/issues/25)
+    * Added check whether `user/accounts` is writable [#25](https://github.com/getgrav/grav-plugin-problems/issues/25)
+    * Added check whether JSON extension is installed
 1. [](#improved)
     * Updated plugin code to the latest standards